Moving game & user controllers in a 'controlers' folder
