+++ lib/application/controllers/game/game-controller.js
@@ -1,5 +1,5 @@
-const Game = require('../../db/models/Game')
-const User = require('../../db/models/User')
+const Game = require('../../../db/models/Game')
+const User = require('../../../db/models/User')
 
 /**
  * Allows to create a payload for the websoacket
