+++ lib/application/controllers/user/user-controller.js
@@ -1,6 +1,6 @@
-const User = require('../../db/models/User')
-const NewUser = require('../../domain/models/NewUser')
-const useCases = require('../../domain/useCases')
+const User = require('../../../db/models/User')
+const NewUser = require('../../../domain/models/NewUser')
+const useCases = require('../../../domain/useCases')
 
 module.exports = {
